Move models.term_similarity, and levenshtein to similarities
