+++ gensim/similarities/docsim.py
@@ -875,7 +875,7 @@ class SoftCosineSimilarity(interfaces.SimilarityABC):
     """
     @deprecated(
         "Method will be removed in 4.0.0, use " +
-        "gensim.models.term_similarity.SparseTermSimilarityMatrix.inner_product instead")
+        "gensim.similarities.termsim.SparseTermSimilarityMatrix.inner_product instead")
     def __init__(self, corpus, similarity_matrix, num_best=None, chunksize=256):
         """
 
@@ -895,7 +895,7 @@ class SoftCosineSimilarity(interfaces.SimilarityABC):
         --------
         :meth:`gensim.models.keyedvectors.WordEmbeddingsKeyedVectors.similarity_matrix`
             A term similarity matrix produced from term embeddings.
-        :func:`gensim.models.levenshtein.similarity_matrix`
+        :func:`gensim.similarities.levenshtein.similarity_matrix`
             A term similarity matrix produced from Levenshtein distances.
         :func:`gensim.matutils.softcossim`
             The Soft Cosine Measure.
@@ -916,13 +916,13 @@ class SoftCosineSimilarity(interfaces.SimilarityABC):
 
     @deprecated(
         "Method will be removed in 4.0.0, use " +
-        "gensim.models.term_similarity.SparseTermSimilarityMatrix.inner_product instead")
+        "gensim.similarities.termsim.SparseTermSimilarityMatrix.inner_product instead")
     def __len__(self):
         return len(self.corpus)
 
     @deprecated(
         "Method will be removed in 4.0.0, use " +
-        "gensim.models.term_similarity.SparseTermSimilarityMatrix.inner_product instead")
+        "gensim.similarities.termsim.SparseTermSimilarityMatrix.inner_product instead")
     def get_similarities(self, query):
         """Get similarity between `query` and current index instance.
 
@@ -969,7 +969,7 @@ class SoftCosineSimilarity(interfaces.SimilarityABC):
 
     @deprecated(
         "Method will be removed in 4.0.0, use " +
-        "gensim.models.term_similarity.SparseTermSimilarityMatrix.inner_product instead")
+        "gensim.similarities.termsim.SparseTermSimilarityMatrix.inner_product instead")
     def __str__(self):
         return "%s<%i docs, %i features>" % (self.__class__.__name__, len(self), self.similarity_matrix.shape[0])
